refactor: 도메인명 변경 및 Menu와 MenuGroup, Order와 OrderTable 패키지 결합
