+++ src/main/java/kitchenpos/menu/application/dto/MenuGroupCreateRequest.java
@@ -1,0 +1,16 @@
+package kitchenpos.menu.application.dto;
+
+public class MenuGroupCreateRequest {
+    private String name;
+
+    private MenuGroupCreateRequest() {
+    }
+
+    public MenuGroupCreateRequest(final String name) {
+        this.name = name;
+    }
+
+    public String getName() {
+        return name;
+    }
+}
