+++ src/main/java/kitchenpos/menu/domain/repository/MenuGroupRepository.java
@@ -1,0 +1,10 @@
+package kitchenpos.menu.domain.repository;
+
+import kitchenpos.menu.domain.MenuGroup;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface MenuGroupRepository extends JpaRepository<MenuGroup, Long> {
+    default MenuGroup getById(final Long id) {
+        return findById(id).orElseThrow(IllegalArgumentException::new);
+    }
+}
