--- src/main/java/kitchenpos/menugroup/application/dto/MenuGroupCreateRequest.java
@@ -1,16 +1,0 @@
-package kitchenpos.menugroup.application.dto;
-
-public class MenuGroupCreateRequest {
-    private String name;
-
-    private MenuGroupCreateRequest() {
-    }
-
-    public MenuGroupCreateRequest(final String name) {
-        this.name = name;
-    }
-
-    public String getName() {
-        return name;
-    }
-}
