--- src/main/java/kitchenpos/menugroup/domain/MenuGroup.java
@@ -1,35 +1,0 @@
-package kitchenpos.menugroup.domain;
-
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.Id;
-
-import static javax.persistence.GenerationType.IDENTITY;
-
-@Entity
-public class MenuGroup {
-    @Id
-    @GeneratedValue(strategy = IDENTITY)
-    private Long id;
-    private String name;
-
-    protected MenuGroup() {
-    }
-
-    public MenuGroup(final Long id, final String name) {
-        this.id = id;
-        this.name = name;
-    }
-
-    public static MenuGroup from(final String name) {
-        return new MenuGroup(null, name);
-    }
-
-    public Long getId() {
-        return id;
-    }
-
-    public String getName() {
-        return name;
-    }
-}
