+++ src/main/java/kitchenpos/order/application/dto/OrderTableId.java
@@ -1,0 +1,16 @@
+package kitchenpos.order.application.dto;
+
+public class OrderTableId {
+    private Long id;
+
+    private OrderTableId() {
+    }
+
+    public OrderTableId(final Long id) {
+        this.id = id;
+    }
+
+    public Long getId() {
+        return id;
+    }
+}
