+++ src/main/java/kitchenpos/order/application/dto/OrderTableResponse.java
@@ -1,0 +1,43 @@
+package kitchenpos.order.application.dto;
+
+import kitchenpos.order.domain.OrderTable;
+
+public class OrderTableResponse {
+
+    private final Long id;
+    private final Long tableGroupId;
+    private final int numberOfGuests;
+    private final boolean empty;
+
+    public OrderTableResponse(Long id, Long tableGroupId, int numberOfGuests, boolean empty) {
+        this.id = id;
+        this.tableGroupId = tableGroupId;
+        this.numberOfGuests = numberOfGuests;
+        this.empty = empty;
+    }
+
+    public static OrderTableResponse from(final OrderTable orderTable) {
+        return new OrderTableResponse(
+                orderTable.getId(),
+                orderTable.getTableGroupId(),
+                orderTable.getNumberOfGuests(),
+                orderTable.isEmpty()
+        );
+    }
+
+    public Long getId() {
+        return id;
+    }
+
+    public Long getTableGroupId() {
+        return tableGroupId;
+    }
+
+    public int getNumberOfGuests() {
+        return numberOfGuests;
+    }
+
+    public boolean isEmpty() {
+        return empty;
+    }
+}
