+++ src/main/java/kitchenpos/order/domain/repository/OrderTableRepository.java
@@ -1,0 +1,14 @@
+package kitchenpos.order.domain.repository;
+
+import kitchenpos.order.domain.OrderTable;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.List;
+
+public interface OrderTableRepository extends JpaRepository<OrderTable,Long> {
+    default OrderTable getById(final Long id) {
+        return findById(id).orElseThrow(IllegalArgumentException::new);
+    }
+
+    List<OrderTable> findAllByIdIn(final List<Long> ids);
+}
