--- src/main/java/kitchenpos/ordertable/application/dto/OrderTableCreateRequest.java
@@ -1,22 +1,0 @@
-package kitchenpos.ordertable.application.dto;
-
-public class OrderTableCreateRequest {
-    private Integer numberOfGuests;
-    private Boolean empty;
-
-    private OrderTableCreateRequest() {
-    }
-
-    public OrderTableCreateRequest(final Integer numberOfGuests, final Boolean empty) {
-        this.numberOfGuests = numberOfGuests;
-        this.empty = empty;
-    }
-
-    public Integer getNumberOfGuests() {
-        return numberOfGuests;
-    }
-
-    public Boolean getEmpty() {
-        return empty;
-    }
-}
