--- src/main/java/kitchenpos/ordertable/application/dto/OrderTableId.java
@@ -1,16 +1,0 @@
-package kitchenpos.ordertable.application.dto;
-
-public class OrderTableId {
-    private Long id;
-
-    private OrderTableId() {
-    }
-
-    public OrderTableId(final Long id) {
-        this.id = id;
-    }
-
-    public Long getId() {
-        return id;
-    }
-}
