--- src/main/java/kitchenpos/ordertable/domain/repository/OrderTableRepository.java
@@ -1,14 +1,0 @@
-package kitchenpos.ordertable.domain.repository;
-
-import kitchenpos.ordertable.domain.OrderTable;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.List;
-
-public interface OrderTableRepository extends JpaRepository<OrderTable,Long> {
-    default OrderTable getById(final Long id) {
-        return findById(id).orElseThrow(IllegalArgumentException::new);
-    }
-
-    List<OrderTable> findAllByIdIn(final List<Long> ids);
-}
