--- src/main/java/kitchenpos/tablegroup/domain/TableGroup.java
@@ -1,48 +1,0 @@
-package kitchenpos.tablegroup.domain;
-
-import kitchenpos.ordertable.domain.OrderTable;
-import kitchenpos.ordertable.domain.OrderTables;
-
-import javax.persistence.Embedded;
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.Id;
-import java.time.LocalDateTime;
-import java.util.List;
-
-import static javax.persistence.GenerationType.IDENTITY;
-
-@Entity
-public class TableGroup {
-    @Id
-    @GeneratedValue(strategy = IDENTITY)
-    private Long id;
-    private LocalDateTime createdDate;
-    @Embedded
-    private OrderTables orderTables;
-
-    protected TableGroup() {
-    }
-
-    public TableGroup(final Long id, final LocalDateTime createdDate, final OrderTables orderTables) {
-        this.id = id;
-        this.createdDate = createdDate;
-        this.orderTables = orderTables;
-    }
-
-    public static TableGroup of(final LocalDateTime now, final List<OrderTable> orderTables) {
-        return new TableGroup(null, now, new OrderTables(orderTables));
-    }
-
-    public Long getId() {
-        return id;
-    }
-
-    public LocalDateTime getCreatedDate() {
-        return createdDate;
-    }
-
-    public OrderTables getOrderTables() {
-        return orderTables;
-    }
-}
