+++ src/test/java/kitchenpos/domain/OrderTableTest.java
@@ -1,6 +1,6 @@
 package kitchenpos.domain;
 
-import kitchenpos.ordertable.domain.OrderTable;
+import kitchenpos.order.domain.OrderTable;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
